Add an image format conversion step into the generate_thumbs task, so any non-PNG images will be converted to PNG.
Had to do this cause Twitter seems to now be converting PNGs to JPGs, so I don't have access to the PNG originals of some images. This isn't the best solution; if a PNG has already been converted to a JPG, converting it back to a PNG isn't gonna restore any quality. In fact it'll probably have a bigger filesize. So we should probably be keeping any JPGs as JPGs, and handling that somehow so that the website serves up the correct image.