Update lightgallery.js repo URL
This comment block's URL was pointing to the jQuery version of
lightGallery, despite being part of the pure JS version. This led me on
a weird journey trying to find the right repo, during which I questioned
my entire existence.

I've opened a PR with the source repo for this change:
https://github.com/sachinchoolur/lightgallery.js/pull/44
but in the meantime I've made it locally here too, so I don't confuse
myself again in future.