Add (most) Gen 7 Pokemon
We got a whole new batch of friends to play with!

KC has started drawing Gen 7 Pokemon but decided to hold off on
possibly "spoilery" Pokemon for the moment [1] so I've gone with the
same approach here. I've added Pokemon up to the point that I would
start to consider their appearance to be story spoilers for Sun & Moon.

[1] Source:
https://twitter.com/kcgreenn/status/809826577366257668
https://twitter.com/kcgreenn/status/809826853691199488
https://twitter.com/kcgreenn/status/809827874773549057