Spike generating silhouette images
For any Pokemon that KC hasn't drawn I'd like to show a washed-out card
with a silhouette of the Pokemon, kinda "Who's That Pokemon???" style.

This script tries to create a silhouette by converting any non-white
pixels to a solid grey. Some blurring is applide to make that work more
nicely.

It works great on most Pokemon, but any with white segments on their
body (or big white eyes) have plenty of white pixels remaining, which
doesn't look great. I'm calling this a non-success.