Add working silhouette gulp task
This is actually 99% the same as the task I just removed, but it relies
on having input images with transparent backgrounds.

The good news of this is that all non-transparent pixels can be reliably
converted to a single colour to create a silhouette.

The bad news is that it means we need to add ~400MB of source images to
the repo. 😫

I could downsize those images before committing them to save on repo
filesize, _but_ it occurs to me that if I'd just committed these
full-size source images in the first place instead of committing
processed downsized images, I wouldn't have had this issue. Having the
biggest possible source images in the best quality format gives us the
widest range of options in future.