Display Sugimori images as Pokemon cards too!
This is rad, we're now displaying all Pokemon as Pokemon cards. Ones
that KC have drawn are colourful (based on the Pokemon's type) and ones
that he hasn't done yet are greyed out with a silhouette.

This code is a bit spikey, it needs to be cleaned up. Also need to
remove leftover unused code for sprite generation.