Remove image alt, get caption from link title
Now that we're lazy loading images, the images don't have a src
attribute until they've been triggered to load by lazyload.js. This
means that their alt text (the Pokemon's name) is shown in the image's
place until the image loads. That looks dodgy. If we remove the alt
text, then we're all good.

This caused another issue though - the Pokemon's name was previously
shown as its caption in the lightbox, but now it had disappeared.
Lightgallery.js will use either the link's title or a child image's alt
text as the caption, so it surprised me that this disappeared since the
Pokemon's name is still present as the link title.

Turns out lightgallery.js' caption logic works like this:
1. if some custom markup (subHtml) has been supplied, use that.
2. if the setting to use title or alt is enabled, and subHtml is blank:
  3. set the caption to the link's title.
  4. if the caption is present and the link has a child image:
    5. set the caption to the alt text on the child image.

It's number 4 that's the problem there; why would you only continue to
look for a caption _if you already had a caption_? This causes a few
possible outcomes:
 - If the link has a title and contains an image with alt text, the alt
   text will be used.
 - If the link has a title and contains an image with no alt text, a
   blank string will be used.
 - If the link has no title and contains an image, the image won't even
   be checked and a blank string will be used.

Surely it's more sensible to use the link title if it's present, and
only continue to look for a caption if we don't already have one? I
certainly think so, so I made that change to lightgallery.js. Feels
dirty modifying vendor code, but I don't really have any other way of
achieving this.