Take out the ableism, replace it with absurdity
I previously thought that referring to KC as "a crazy person" was
harmlessly funny, but the more I think about it the less I think that.
It's pretty inappropriate and derogatory.

So while I'm coming up with something new to put here, why not come up
with SEVERAL new things to put here, and randomly select one each time
the page loads? Obviously this is a normal thing to do.