Minify HTML for excellent filesize savings
Adds `gulp-htmlmin` to easily remove a whooole bunch of extraneous
whitespace and unnecessary quotation marks.

In addition to that, I've wrapped a few `data-` attributes in an
`if data.has_kc_image` conditional, since they're only used by our JS
when the Pokemon has a KC image. More savings for the other Pokemons!

All up that brings the size of index.html down from 242KB to 132KB,
which is a 45% reduction! WOOOOOP