Stop busting the cache of site images and fonts
These files rarely change, but we're creating new versions of them
every time we build the site. That means we're busting the cache of
those files for users who have already downloaded them, despite the fact
that the files' contents are identical.

This adds `gulp-newer` to the gulp tasks that create those files, so
that we're no longer recreating them unless we need to.