Add an .htaccess file to cache images super hard
I'm hosting this site on pretty crappy shared hosting, and trying to
load roughly 800 images from the site at once leads to a lot of failures
because the server just doesn't handle it. We've got Cloudflare in front
of the server acting as a CDN, but we're only telling Cloudflare to
cache these images for two hours, which is barely anything.

These images should never change, so we could (and should!) really cache
them for as long as possible. This adds a .htaccess file which tells
browsers and Cloudflare to cache all images for a year. That oughtta
give the server a break!