Don't choose same random Pokemon twice in a row
When hitting the "Gimme another" button, the randomizer sometimes
chooses the same Pokemon that is currently displayed. This change forces
it to try again until it selects a different Pokemon.

That does create a fun infinite loop if there's only one Pokemon left,
so let's not display the button in that case. Let's display a special
congratulatory message for KC instead.