Make sure randomizer can display all options
Shit! I had a bug where the last Pokemon in the list of undrawn Pokemon
would never appear. That's bad! Kind of important to have that work!

I've tested this a bit more thoroughly now and can confirm that it
works. All remaining Pokemon can be selected by the randomizer.