Set black background for Haunter and Shiinotic
These two illustrations have a black background, but their thumbmails
get a white border added around them just like every other thumbnail,
which looks gross.

I've wanted to fix this for ages, but couldn't find a way to automate it
(e.g. detect the background colour from the source image, and use that
when adding a border). It didn't occur to me for ages that I could just
add some special case code for these two images instead. It's not the
nicest code, but it looks SO MUCH BETTER ✨