Don't flash broken thumbnails for unloaded images
Sometimes our server gets overwhelmed with the amount of images being
loaded at once, and some of the images fail to load. We detect this with
JavaScript and attempt to load the image again until it's successful,
but this causes a broken image icon to briefly flash onto the page.
That's a little janky, so let's just hide the image using CSS until it's
been loaded successfully. We show a loading animation behind the image
anyway, so this should provide a nicer user experience.