Upgrade gulp-autoprefixer package
This new version throws up a big red warning that we should move our
browser config into package.json so that it can be shared with other
tools that might use it. We don't have any other tools that will use it
in this repo, but sure. Why not.

This still seems to work nicely - in the generated output I can see
there are way fewer -webkit prefixes, which suggests those are no longer
necessary in 2020 and our 'last 5 versions' config is working!