Upgrade gulp-imagemin package
That's a lotta major version jumps! But it's almost entirely due to
internal dependency changes (for different versions of the imagemin
plugins it uses, and different versions of Node).

The only major change for us happened in 3.0.0, where instead of passing
a hash of options directly to imagemin, we specify which imagemin
plugins we want to use and pass options to each of them. Details:
https://github.com/sindresorhus/gulp-imagemin/releases/tag/v3.0.0

I've switched to passing no options now for a couple reasons:
- The defaults options are pretty good, and give us about the same level
  of optimisation we were previously getting
- Specifying options requires listing out all plugins we want imagemin
  to use, which means we're duplicating its defaults and would need to
  keep that list up-to-date when doing future upgrades
- I tried increasing the optimisation settings for the mozjpeg and
  optipng plugins a little bit, and it massively increased the build
  time for a full rebuild (from 7.5 minutes to 21 minutes! 😵) while
  only saving us an extra 800KB (2%)

So yeah, let's stick with the defaults! Keeps the code cleaner anyway.