Replace gulp-minify-css with gulp-clean-css
I went to upgrade gulp-minify-css and discovered it's been marked as
deprecated, with the author recommending to use gulp-clean-css instead.
They're both wrappers around the clean-css package, so meh. Either or.

After this upgrade we're producing _slightly_ different CSS, but nothing
weird. The major differences are that all hex colour codes are being
downcased now, and not as many selectors are being combined to share
the same set of rules (e.g. everything that sets `display: none` used to
get crammed together into one selector, but now they don't). That's
probably good, it'll make things easier to debug anyway.