Upgrade gulp-newer package, use it for more tasks
Not any public changes to gulp-newer, but it did have a dependency on
gulp-util that's been removed. One less thing.

While updating it, I noticed in its readme that it has the ability to
handle many-to-one source-to-destination mappings. Currently we've only
been using it for one-to-one checking (e.g. is 123.jpg in the source
folder newer than 123.jpg in the build folder?), and for anything that
combines multiple files into one (e.g. JS, CSS) we've been straight-up
deleting all destination files on every rebuild. This changes our setup
to use that many-to-one feature on our JS and CSS, so we only rebuild
scripts and styles if something has changed. Works nicely!