Upgrade gulp-uglify-js package, remove fancy-log
gulp-uglify-js has some better built-in error reporting now (or maybe
it's just generally not as cryptic in gulp 4?) so we don't need to pass
fancy-log to it for extra error logging anymore. Which means we don't
need fancy-log anymore!

gulp-uglify-js does recommend adding some extra stuff for error logging,
but it's more stuff to install and I'm pretty happy with what we get out
of the box so I'm skipping it!

This gets rid of our last dependency on gulp-util! It's gone! We're now
officially using gulp 4 mostly as intended! Exclamation marks!