Ensure clean tasks returns as expected
gulp 4 expects tasks to be async-friendly, and needs tasks to do some
special handling if they run synchronous code. If your task is
synchronous but doesn't do the special handling, gulp prints a warning:

```
$ gulp clean
[17:29:36] Using gulpfile ~/Dev/your-pokemon-for-today/gulpfile.js
[17:29:36] Starting 'clean'...
[17:29:37] The following tasks did not complete: clean
[17:29:37] Did you forget to signal async completion?
```

Looks like the solution is to use `del()` instead of `del.sync()`.
`del()` returns a Promise, which gulp is happy with.