Give every Pokemon a "dex_number" attribute
I'm trying to add better support for Pokemon with regional variant
forms, which means I need to display multiple cards that share the same
dex number. That means the dex number is no longer a useful unique ID,
so I need to split out the dex number as a separate piece of data.

This splits out the `dex_number` for each Pokemon as a separate
attribute, alongside their name and types. Now the key pointing to their
data will be their unique ID, and the dex number is just a piece of data
that doesn't have to be unique. For the moment the unique ID and the dex
number are the same for all Pokemon though. That'll change soon!