Add data entry and image for Alolan Raichu
Here's the first test of whether this new system will work: adding a
regional variant form for Alolan Raichu, which shares the same dex
number as Raichu. Currently this will probably just break stuff, but the
next step is to get this showing up as a separate card next to Raichu.