Add Galarian Slowpoke data, split Raichu images
Testing this regional variant stuff, I quickly discovered it'd be handy
to have an example of a regional variant with a KC image, and another
without a KC image. This splits out Raichu's image into two so that can
be the former, and adds data for Galarian Slowpoke so that can be the
latter.