Update build code to handle new data format
The shape of our source data has changed - check out the previous few
commits for details on that. This updates the build code to handle the
new data, and the fact that source image filenames can now contain more
characters than just digits.

Some of these changes aren't strictly necessary - they just update some
variable names to reflect the fact that the "key" we refer to in some
places is now the unique ID that identifies a Pokemon form.

This is mostly working! The regional variant forms are appearing on the
compiled site, but they're appearing at the end of the page instead of
next to their counterpart forms. Gotta suss out some sorting logic next
then.