Restore original images for both of KC's Raichus
When I first added Alolan Raichu, I smushed it into the same image as
Kantonian Raichu, and probably resized one or both of them. Then to
experiment with showing them as separate images, I cut them out of that
shared image and moved them into their own files. That would've kept
whatever downscaling I applied to them when I hacked them together,
which sucks!

This restores the original images for both of them. For Kantonian Raichu
I could grab the original from Git history, but Alolan Raichu never made
it into Git history so I had to find the original on Twitter. That was
not fun! The only reference I had for where to look was the date I added
it to this repo, and it ended up being posted a few weeks before that.
Twitter's search didn't make it easy to find it, but I did it! Not
looking forward to doing the rest of the ones I've gotta split up.