Add "hasha" package for hashing files
Currently filenames for images, CSS/JS files etc. stay the same across
builds, which means that if the file contents change, I need to manually
bust the CDN cache for each changed file in order to get the new
versions served to visitors. But even that doesn't always work, because
return visitors might have the file cached locally, in which case
they'll likely never see the new version. Hmm. Crap.

I'd like to start adding a hash to the filename of each built file,
based on the contents of the source file. This way the filename will
change when the file contents change, and no cache-busting will be
required. This change imports the "hasha" package which is able to
generate a hash for a given file: https://www.npmjs.com/package/hasha