Add function to generate a hash for a given file
This will let us append a hash to the filename of each image, based on
the contents of the source file, so that the output image's filename
will change when the contents of the source file change. This means
images will be cacheable forever, since their filename will change any
time their contents change.

This doesn't do the whole thing yet - just sets up a function for
getting the hash for a file (8 characters oughtta be enough I reckon!),
and prints out the hash for a bunch of files at build time. Next step is
to use those hashes somehow!

And yes there are some differences in code style here. It's 2020 baby!
I'd been sticking to the style that was already in use in this file
because I didn't want to update the whole file, but that's not a good
reason to keep adding new code in the old style. Let's update things a
little bit at a time.