Add hashes to KC thumb image filenames
Hmm there's a lot to this. We have to calculate the destination filename
twice - once for `newer` to check if the source file is newer than the
destination file, and once for `rename` to rename the destination file
after it's been processed. We can't even really pull that out into a
shared function, since those two packages require the destination
filename in different formats.

But! It's working! This is now creating thumbs with hashed filenames,
and then looking for the hashed version when doing the comparison to see
if the thumbnail needs to be regenerated.

However this isn't piping those hashed filenames through to the HTML
page that gets generated, so that's still pointing at the unhashed
files. We'll get there soon.