Add hashes to Sugimori thumb image filenames
Copying the logic we used for KC thumbnails to the official art images!
Turns out the function we use to get the destination filename is the
same as what we need for the KC images, so I've renamed that function to
be more generic so we can reuse it here.

Again, these hashed destination filenames aren't being used by the HTML
page yet, so they go unused on the site for now.