Use hashed Sugimori image filenames on the site
Woo! Site is now using the immutable filenames, which should eliminate
caching issues.

These ones took a little bit of extra setup, because we were (ab)using
the `thumb_filepath` property later on by using it as the source
filepath for the image when it is only intended to be the destination
filepath. That's cool, we fixed that by adding a `source_image_filepath`
property to be used for that purpose. This makes everything way more
legit!