Use a hashed filename for generated JS file
Hoo, a lot going on here! This is adopting what we've just done for all
of the images, taking a hash of the contents of the source file and
using that as part of the filename of the destination file, so that the
destination filename changes when the source files change, automatically
busting its own cache.

With the images we had a 1:1 mapping between each destination file and a
source file, but with the JS file we compile multiple source files into
a single destination file. So we need to calculate a single hash based
on the contents of all of those multiple source files. Here's how we do
that:

- Take the list of glob patterns we've got in `paths.scripts` that is
  used to figure out which source files to combine into a single
  destination file, and figure out which files that'll evaluate to. This
  required us to pull in the `glob` package, which was already installed
  since it's used by a bunch of our Gulp plugins, but wasn't explicitly
  listed as a dependency because we weren't using it directly.
- Calculate a hash of the contents of each of those files, sorting the
  files by filename first to make sure we always have the list in the
  same order.
- Calculate a hash of the list of hashes for those files, and use that
  as the hash in our destination filename!

This works pretty nicely! However we need to tell our HTML files what
the hash is. Previously they were using the app's version number as part
of the JS filename, and we're doing the same for our CSS file (which
we'll apply this logic to next!).

Yes I realise I'm kinda building half of Webpack here. But we don't need
the main functionality of Webpack, and this is more fun anyway.