Use a hashed filename for generated CSS file
Super easy! Just gotta remember to include those `styles_extra` files.

An unexpected side benefit of this change is that the CSS and JS files
are no longer sharing the app version as a version identifier, so making
changes to JS won't cause the CSS file's cache to be busted (and vice
versa).